Terrain scene
I created the basic terrain. I'm not sure what to do with the view of infinite nothingness you can see at either end of the canyon, though I have a few ideas. I would've just made the terrain larger, but I thought it might be a bit big already.  The fort plateau may or may not be low as well.